Added commands & renamed
