+++ src/Files.App/ViewModels/UserControls/Widgets/DrivesWidgetViewModel.cs
@@ -180,7 +180,6 @@ namespace Files.App.ViewModels.UserControls.Widgets
 					Text = "Loading".GetLocalizedResource(),
 					Glyph = "\xE712",
 					Items = [],
-					ID = "ItemOverflow",
 					Tag = "ItemOverflow",
 					IsEnabled = false,
 				}
